fix selected plot mode conflict with edited mode
