+++ frontend/src/renderer/components/grid/HoverButtons.tsx
@@ -18,7 +18,11 @@ import {
 } from '@tabler/icons-react';
 import { useHover } from '@mantine/hooks';
 import { Configuration, DataGridPlot } from '../../types';
-import { applyRange, fetchErrorBandsInConfig } from '../../utils';
+import {
+  applyRange,
+  containsFloat,
+  fetchErrorBandsInConfig,
+} from '../../utils';
 import { useIbexStore } from '../../stores';
 
 interface HoverButtonsProps {
@@ -29,7 +33,6 @@ interface HoverButtonsProps {
   handleCustomization: (id: string) => void;
   handleDeleteGrid: (id: string) => void;
   is3DView: boolean;
-  setIs3DView: React.Dispatch<React.SetStateAction<boolean>>;
   active3DTab: string;
   setActive3DTab: React.Dispatch<React.SetStateAction<string>>;
 }
@@ -43,7 +46,6 @@ export const HoverButtons = React.memo(
     handleCustomization,
     handleDeleteGrid,
     is3DView,
-    setIs3DView,
     active3DTab,
     setActive3DTab,
   }: HoverButtonsProps) => {
@@ -159,6 +161,35 @@ export const HoverButtons = React.memo(
       updateErrorBands();
     }, [data.displayErrorBand]);
 
+    const updateSelectedPlotMode = (
+      is3DView: boolean,
+      active: Configuration,
+    ) => {
+      const updatedDataPlot: DataGridPlot[] = JSON.parse(
+        JSON.stringify(active.dataPlot),
+      );
+      const selectedDataPlot = updatedDataPlot.find(
+        (dataPlot) => dataPlot.i === data.i,
+      );
+      if (selectedDataPlot?.selectedPlotMode) {
+        selectedDataPlot.selectedPlotMode = is3DView ? 'Heatmap' : '1D';
+      } else {
+        selectedDataPlot.selectedPlotMode =
+          data.coordinates.length >= 2 &&
+          containsFloat(
+            data.coordinates.find((coord) => coord.axeIndex === 1)?.data,
+          )
+            ? 'Heatmap'
+            : '1D';
+      }
+
+      const updatedActive: Configuration = {
+        ...active,
+        dataPlot: updatedDataPlot,
+      };
+      updatedConfiguration(updatedActive);
+    };
+
     return (
       <div ref={hoverRef} className={classes.containerButton}>
         <Group justify="space-between" h={'100%'}>
@@ -215,7 +246,12 @@ export const HoverButtons = React.memo(
                   <ActionIcon
                     variant="filled"
                     aria-label="Toggle 1D/Heatmap view"
-                    onClick={() => setIs3DView((prev) => !prev)}
+                    onClick={() =>
+                      updateSelectedPlotMode(
+                        !(data.selectedPlotMode === 'Heatmap'),
+                        active,
+                      )
+                    }
                     className={classes.actionButton}
                   >
                     {is3DView ? <Text fw="bold">1D</Text> : heatmapLogo}
@@ -262,11 +298,7 @@ export const HoverButtons = React.memo(
               )}
 
               <Tooltip
-                label={
-                  data.isEditing
-                    ? 'Validate/Close editing the grid'
-                    : 'Open editing the grid'
-                }
+                label={data.isEditing ? 'Save the edition' : 'Edit the grid'}
               >
                 <ActionIcon
                   variant="filled"
