+++ frontend/src/renderer/utils/matrix.ts
@@ -43,7 +43,9 @@ export const getArrayValueFromDependance = (
   }
 
   // get sorted valueIndex list (sorted by shape length) to access to data matrix
-  const dependances = JSON.parse(JSON.stringify(wantedCoordinate.coord_dependencies)).reverse();
+  const dependances = JSON.parse(
+    JSON.stringify(wantedCoordinate.coord_dependencies),
+  ).reverse();
   const sortedIndexValueDependances: number[] = [];
   for (const dependance of dependances) {
     const coordDep = coordinates.find(
